+++ Tests/Unit/ClientTest.php
@@ -5,6 +5,7 @@ namespace F3\CouchDB;
  * A rather functional test for the CouchDB client
  */
 class ClientTest extends \F3\Testing\BaseTestCase {
+
 	/**
 	 * @var \F3\CouchDB\Client
 	 */
@@ -14,11 +15,27 @@ class ClientTest extends \F3\Testing\BaseTestCase {
 	 * Setup a CouchDB HTTP connector
 	 */
 	public function setUp() {
-		$this->client = new \F3\CouchDB\Client('http://127.0.0.1:5984');
+		$connector = new \F3\CouchDB\Client\HttpConnector('127.0.0.1', '5984');
+		$mockObjectManager = $this->getMock('F3\FLOW3\Object\ObjectManagerInterface');
+		$mockObjectManager->expects($this->once())->method('create')->with('F3\CouchDB\Client\HttpConnector', '127.0.0.1', '5984', NULL, NULL, array())->will($this->returnValue($connector));
+		$this->client = $this->getAccessibleMock('F3\CouchDB\Client', array('dummy'), array('http://127.0.0.1:5984'));
+		$this->client->_set('objectManager', $mockObjectManager);
+		$this->client->initializeObject();
 
-		$this->client->deleteDatabase('flow3_test');
+		if ($this->client->databaseExists('flow3_test')) {
+			$this->client->deleteDatabase('flow3_test');
+		}
 		$this->client->createDatabase('flow3_test');
-		$this->client->setDatabase('flow3_test');
+		$this->client->setDatabaseName('flow3_test');
+	}
+
+	/**
+	 * Remove flow3_test database
+	 */
+	public function tearDown() {
+		if ($this->client->databaseExists('flow3_test')) {
+			$this->client->deleteDatabase('flow3_test');
+		}
 	}
 
 	/**
@@ -41,9 +58,9 @@ class ClientTest extends \F3\Testing\BaseTestCase {
 	 * @test
 	 */
 	public function createDocumentWithIdWorks() {
-		$response = $this->client->createDocument('abc', array(
+		$response = $this->client->createDocument(array(
 			'name' => 'Foo'
-		));
+		), 'abc');
 		$this->assertTrue($response->isSuccess());
 	}
 
@@ -76,13 +93,13 @@ class ClientTest extends \F3\Testing\BaseTestCase {
 	 * @test
 	 */
 	public function updateDocumentWorks() {
-		$response = $this->client->createDocument('abc', array(
+		$response = $this->client->createDocument(array(
 			'name' => 'Foo'
-		));
-		$response = $this->client->updateDocument('abc', array(
+		), 'abc');
+		$response = $this->client->updateDocument(array(
 			'name' => 'Bar',
 			'_rev' => $response->getRevision()
-		));
+		), 'abc');
 		$this->assertTrue($response->isSuccess());
 
 		$response = $this->client->getDocument('abc');
@@ -93,9 +110,9 @@ class ClientTest extends \F3\Testing\BaseTestCase {
 	 * @test
 	 */
 	public function deleteDocumentWorks() {
-		$response = $this->client->createDocument('abc', array(
+		$response = $this->client->createDocument(array(
 			'name' => 'Foo'
-		));
+		), 'abc');
 		$response = $this->client->deleteDocument('abc', $response->getRevision());
 		$this->assertTrue($response);
 
@@ -111,9 +128,9 @@ class ClientTest extends \F3\Testing\BaseTestCase {
 	 * @test
 	 */
 	public function listDocumentsWithParameters() {
-		$this->client->createDocument('abc', array(
+		$this->client->createDocument(array(
 			'name' => 'Foo'
-		));
+		), 'abc');
 
 		$response = $this->client->listDocuments(array('include_docs' => TRUE));
 		$this->assertEquals(1, count($response->rows));
@@ -123,15 +140,15 @@ class ClientTest extends \F3\Testing\BaseTestCase {
 	 * @test
 	 */
 	public function getDocumentsWithParameters() {
-		$this->client->createDocument('abc', array(
+		$this->client->createDocument(array(
 			'name' => 'Foo'
-		));
-		$this->client->createDocument('def', array(
+		), 'abc');
+		$this->client->createDocument(array(
 			'name' => 'Bar'
-		));
-		$this->client->createDocument('ghi', array(
+		), 'def');
+		$this->client->createDocument(array(
 			'name' => 'Baz'
-		));
+		), 'ghi');
 
 		$response = $this->client->getDocuments(array('abc', 'ghi'), array('include_docs' => TRUE));
 		$this->assertEquals(2, count($response->rows));
@@ -141,9 +158,9 @@ class ClientTest extends \F3\Testing\BaseTestCase {
 	 * @test
 	 */
 	public function getDocumentWorks() {
-		$this->client->createDocument('abc', array(
+		$this->client->createDocument(array(
 			'name' => 'Foo'
-		));
+		), 'abc');
 
 		$response = $this->client->getDocument('abc');
 		$this->assertEquals('Foo', $response->name);
@@ -153,7 +170,7 @@ class ClientTest extends \F3\Testing\BaseTestCase {
 	 * @test
 	 */
 	public function createAndGetDesignDocumentWorks() {
-		$this->client->createDocument('_design/test', array('language' => 'javascript'));
+		$this->client->createDocument(array('language' => 'javascript'), '_design/test');
 
 		$response = $this->client->getDocument('_design/test');
 
@@ -171,14 +188,14 @@ class ClientTest extends \F3\Testing\BaseTestCase {
 			'name' => 'Bar'
 		));
 
-		$this->client->createDocument('_design/test', array(
+		$this->client->createDocument(array(
 			'language' => 'javascript',
 			'views' => array(
 				'byName' => array(
 					'map' => 'function(doc) { if (doc.name) { emit(doc.name, null); } }'
 				)
 			)
-		));
+		), '_design/test');
 
 		$response = $this->client->queryView('test', 'byName', array('key' => 'Foo', 'include_docs' => TRUE));
 		$this->assertEquals(1, count($response->rows));
@@ -201,14 +218,14 @@ class ClientTest extends \F3\Testing\BaseTestCase {
 			'name' => 'Baz'
 		));
 
-		$this->client->createDocument('_design/test', array(
+		$this->client->createDocument(array(
 			'language' => 'javascript',
 			'views' => array(
 				'byName' => array(
 					'map' => 'function(doc) { if (doc.name) { emit(doc.name, null); } }'
 				)
 			)
-		));
+		), '_design/test');
 
 		$response = $this->client->queryView('test', 'byName', array('keys' => array('Foo', 'Baz'), 'include_docs' => TRUE));
 		$this->assertEquals(2, count($response->rows));
